Just switching the table to be *const gave me:
    unable to perform tail call: type of function being called '{}' does not match type of calling function '{}'

The problem that seems to give with 0.10 stage2 is that I had been inlining
the dispatch call, but also calling that with .always_tail. This was nice
in that I could easily switch between inline dispatch and noinline dispatch
to compare the two. With the type check that now fails, because the inline
dispatch function type includes the fact that its calling convention is .Noinline
which is not _exactly_ the same as all of the instruction handlers.

To get around this for the moment I will just opt for only using inline.
I then need to just directly call dispatch (not with @call .always_tail).

The reason I can't now quickly check out switching dispatch to noinline
is that it will work yes, but it will cause stack frame allocations
which we absolutely do not want.

If Sema remains unchanged, one we might handle both is by having some
compile time flag that can be set that will directly call inline dispatch
in the case of wanting inline, and calling a noinline dispatch via @call
.always_tail if that's what we want.

The other alternative is that Sema is wrong and that instead of comparing
the complete function type (including calling convention) it should only
compare arg types and return types, in other words don't consider the
calling convention